+++ base/exports.jl
@@ -638,6 +638,7 @@ export
     split,
     string,
     strip,
+    takestring!,
     textwidth,
     thisind,
     titlecase,